converter/requirements.txt
cupy-cuda12x==13.3.0
numpy==2.0.2
fastrlock==0.8.3
PyMuPDF==1.25.1
GPUtil==1.4.0
Pillow==11.1.0
Werkzeug==3.1.3
setuptools==75.8.0